requirements.txt
asgiref==3.4.1
attrs==21.4.0
autopep8==1.6.0
az-iranian-bank-gateways==1.6.13
beautifulsoup4==4.10.0
beautifultable==0.8.0
bleach==4.1.0
cached-property==1.5.2
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.6
click==8.0.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
dj-database-url==0.5.0
Django==3.2.7
django-admin-interface==0.17.3
django-autoslug==1.9.8
django-block-ip==0.2.0
django-ckeditor==6.1.0
django-colorfield==0.4.5
django-cors-headers==3.10.1
django-datetimepicker==3.14
django-filter==2.4.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-formtools==2.3
django-jalali==4.3.0
django-jalali-date==0.3.2
django-js-asset==1.2.2
django-nine==0.2.5
django-nonefield==0.4
django-recaptcha==3.0.0
django-rest-knox==4.1.0
django-taggit==1.5.1
django-user-activity-log2==0.0.24
django-utils-six==2.0
djangorestframework==3.12.4
easy-thumbnails==2.7.1
fandogh-cli==1.26.10
future==0.15.0
gunicorn==20.1.0
idna==3.2
ipcalc==1.99.0
ipython-genutils==0.2.0
isodate==0.6.1
itypes==1.2.0
jdatetime==3.6.2
Jinja2==3.0.3
jsonschema==4.3.2
jupyter-core==4.9.1
lxml==4.7.1
Markdown==3.3.4
MarkupSafe==2.0.1
nbformat==5.1.3
packaging==21.0
Pillow==8.3.1
platformdirs==2.4.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycodestyle==2.8.0
pycparser==2.20
pycryptodome==3.12.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-gnupg==0.4.7
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-file==1.5.1
requests-toolbelt==0.9.1
retrying==1.3.3
six==1.16.0
soupsieve==2.3.1
sphinx-me==0.3
sqlparse==0.4.1
toml==0.10.2
traitlets==5.1.1
tzlocal==2.1
Unidecode==1.2.0
uritemplate==4.1.1
urllib3==1.26.6
vishap==0.1.5
webencodings==0.5.1
websocket-client==1.2.1
whitenoise==5.3.0
xlwt==1.3.0
zeep==4.1.0
